move image converter to gin
