+++ filenames.gni
@@ -468,7 +468,8 @@ filenames = {
     "shell/common/gin_converters/file_dialog_converter_gin_adapter.h",
     "shell/common/gin_converters/file_path_converter.h",
     "shell/common/gin_converters/gurl_converter.h",
-    "shell/common/gin_converters/image_converter_gin_adapter.h",
+    "shell/common/gin_converters/image_converter.cc",
+    "shell/common/gin_converters/image_converter.h",
     "shell/common/gin_converters/message_box_converter.cc",
     "shell/common/gin_converters/message_box_converter.h",
     "shell/common/gin_converters/native_window_converter.h",
@@ -500,7 +501,6 @@ filenames = {
     "shell/common/native_mate_converters/gfx_converter.cc",
     "shell/common/native_mate_converters/gfx_converter.h",
     "shell/common/native_mate_converters/gurl_converter.h",
-    "shell/common/native_mate_converters/image_converter.cc",
     "shell/common/native_mate_converters/image_converter.h",
     "shell/common/native_mate_converters/map_converter.h",
     "shell/common/native_mate_converters/native_window_converter.h",
