+++ shell/common/gin_converters/image_converter.cc
@@ -1,14 +1,14 @@
-// Copyright (c) 2014 GitHub, Inc.
+// Copyright (c) 2019 GitHub, Inc.
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/common/native_mate_converters/image_converter.h"
+#include "shell/common/gin_converters/image_converter.h"
 
 #include "shell/common/api/atom_api_native_image.h"
-#include "shell/common/native_mate_converters/file_path_converter.h"
+#include "shell/common/gin_converters/file_path_converter.h"
 #include "ui/gfx/image/image_skia.h"
 
-namespace mate {
+namespace gin {
 
 bool Converter<gfx::ImageSkia>::FromV8(v8::Isolate* isolate,
                                        v8::Local<v8::Value> val,
@@ -27,8 +27,9 @@ bool Converter<gfx::Image>::FromV8(v8::Isolate* isolate,
   if (val->IsNull())
     return true;
 
-  Handle<electron::api::NativeImage> native_image;
-  if (!ConvertFromV8(isolate, val, &native_image))
+  // TODO(deermichel): remove mate:: after dropping mate
+  mate::Handle<electron::api::NativeImage> native_image;
+  if (!mate::ConvertFromV8(isolate, val, &native_image))
     return false;
 
   *out = native_image->image();
@@ -40,4 +41,4 @@ v8::Local<v8::Value> Converter<gfx::Image>::ToV8(v8::Isolate* isolate,
   return ConvertToV8(isolate, electron::api::NativeImage::Create(isolate, val));
 }
 
-}  // namespace mate
+}  // namespace gin
