+++ shell/common/gin_converters/message_box_converter.cc
@@ -5,7 +5,7 @@
 #include "shell/common/gin_converters/message_box_converter.h"
 
 #include "gin/dictionary.h"
-#include "shell/common/gin_converters/image_converter_gin_adapter.h"
+#include "shell/common/gin_converters/image_converter.h"
 #include "shell/common/gin_converters/native_window_converter.h"
 
 namespace gin {
