+++ shell/common/native_mate_converters/image_converter.h
@@ -6,11 +6,7 @@
 #define SHELL_COMMON_NATIVE_MATE_CONVERTERS_IMAGE_CONVERTER_H_
 
 #include "native_mate/converter.h"
-
-namespace gfx {
-class Image;
-class ImageSkia;
-}  // namespace gfx
+#include "shell/common/gin_converters/image_converter.h"
 
 namespace mate {
 
@@ -18,15 +14,22 @@ template <>
 struct Converter<gfx::ImageSkia> {
   static bool FromV8(v8::Isolate* isolate,
                      v8::Local<v8::Value> val,
-                     gfx::ImageSkia* out);
+                     gfx::ImageSkia* out) {
+    return gin::ConvertFromV8(isolate, val, out);
+  }
 };
 
 template <>
 struct Converter<gfx::Image> {
   static bool FromV8(v8::Isolate* isolate,
                      v8::Local<v8::Value> val,
-                     gfx::Image* out);
-  static v8::Local<v8::Value> ToV8(v8::Isolate* isolate, const gfx::Image& val);
+                     gfx::Image* out) {
+    return gin::ConvertFromV8(isolate, val, out);
+  }
+  static v8::Local<v8::Value> ToV8(v8::Isolate* isolate,
+                                   const gfx::Image& val) {
+    return gin::ConvertToV8(isolate, val);
+  }
 };
 
 }  // namespace mate
